+++ database/factories/TrackActionFactory.php
@@ -26,7 +26,7 @@ class TrackActionFactory extends Factory
     {
         return [
             'action' => $this->faker->randomElement($this->randomAction()),
-            'model_table' => 'people',
+            'model_table' => $this->faker->randomElement(['people', 'address']),
             'model_key' => $this->faker->randomNumber(3),
             'model_changes' => $this->faker->randomElements($this->modelChanges()),
         ];